+++ Cavern.QuickEQ/EQCurves/RoomCurve.cs
@@ -79,7 +79,7 @@ namespace Cavern.QuickEQ.EQCurves {
         /// <param name="length">Curve length</param>
         /// <param name="startFreq">Frequency at the beginning of the curve</param>
         /// <param name="endFreq">Frequency at the end of the curve</param>
-        public override float[] GenerateLogCurve(int length, float startFreq, float endFreq) {
+        public override float[] GenerateLogCurve(int length, double startFreq, double endFreq) {
             float[] curve = new float[length];
             double powerMin = Math.Log10(startFreq), powerRange = length / (Math.Log10(endFreq) - powerMin);
             int lowKnee = (int)((log10_200 - powerMin) * powerRange), highKnee = (int)((log10_1000 - powerMin) * powerRange);
@@ -106,8 +106,8 @@ namespace Cavern.QuickEQ.EQCurves {
         /// <param name="startFreq">Frequency at the beginning of the curve</param>
         /// <param name="endFreq">Frequency at the end of the curve</param>
         /// <param name="gain">Curve reference level</param>
-        /// <remarks>For uses where gain is not needed, use <see cref="GenerateLogCurve(int, float, float)"/>, it's faster.</remarks>
-        public override float[] GenerateLogCurve(int length, float startFreq, float endFreq, float gain) {
+        /// <remarks>For uses where gain is not needed, use <see cref="GenerateLogCurve(int, double, double)"/>, it's faster.</remarks>
+        public override float[] GenerateLogCurve(int length, double startFreq, double endFreq, float gain) {
             float[] curve = new float[length];
             double powerMin = Math.Log10(startFreq), powerRange = length / (Math.Log10(endFreq) - powerMin);
             int lowKnee = (int)((log10_200 - powerMin) * powerRange), highKnee = (int)((log10_1000 - powerMin) * powerRange);
